Remove circular references and move SPDX licences to field model
