+++ amati/fields/__init__.py
@@ -8,5 +8,4 @@ Convenience, can import all types from one place
 from amati.fields._custom_types import _Str
 from amati.fields.http_status_codes import HTTPStatusCode
 from amati.fields.media import MediaType
-from amati.fields.reference import Reference, References
 from amati.fields.uri import URI
