+++ tests/fields/test_spdx_licences.py
@@ -2,17 +2,18 @@
 Tests amati/fields/spdx_licences.py
 """
 
+import pytest
 from hypothesis import given
 from hypothesis import strategies as st
 from hypothesis.provisional import urls
 
+from amati import AmatiValueError
 from amati.fields.spdx_licences import (
     SPDXURL,
     VALID_LICENCES,
     VALID_URLS,
     SPDXIdentifier,
 )
-from amati.logging import LogMixin
 from amati.validators.generic import GenericObject
 
 VALID_IDENTIFIERS = list(VALID_LICENCES.keys())
@@ -30,32 +31,23 @@ class URLModel(GenericObject):
 
 
 @given(st.sampled_from(VALID_IDENTIFIERS))
-def test_spdx_identifier_valid(identifier: str):
-    IdentifierModel(identifier=identifier)
-    print(LogMixin.logs)
+def test_spdx_identifier_valid(value: str):
+    SPDXIdentifier(value)
 
 
 @given(st.text())
-def test_spdx_identifier_invalid(identifier: str):
-    with LogMixin.context():
-        IdentifierModel(identifier=identifier)
-        assert LogMixin.logs
-        assert LogMixin.logs[0].message is not None
-        assert LogMixin.logs[0].type == Warning
-        assert LogMixin.logs[0].reference is not None
+def test_spdx_identifier_invalid(value: str):
+    with pytest.raises(AmatiValueError):
+        SPDXIdentifier(value)
 
 
 @given(st.sampled_from(VALID_URLS))
-def test_spdx_url_valid(url: str):
+def test_spdx_url_valid(value: str):
     # Expecting that the URL is passed as a string from JSON
-    URLModel(url=url)  # type: ignore
+    SPDXURL(value)
 
 
 @given(urls())
-def test_spdx_url_invalid(url: str):
-    with LogMixin.context():
-        URLModel(url=url)  # type: ignore
-        assert LogMixin.logs
-        assert LogMixin.logs[0].message is not None
-        assert LogMixin.logs[0].type == Warning
-        assert LogMixin.logs[0].reference is not None
+def test_spdx_url_invalid(value: str):
+    with pytest.raises(AmatiValueError):
+        SPDXURL(value)
