+++ tests/validators/test_licence_object.py
@@ -27,9 +27,8 @@ def test_all_variables_correct(name: str, identifier: str):
 
 @given(helpers.text_excluding_empty_string(), INVALID_IDENTIFIERS, INVALID_URLS)
 def test_all_variables_random(name: str, identifier: str, url: str):
-    with LogMixin.context():
+    with pytest.raises(ValidationError):
         LicenceObject(name=name, identifier=identifier, url=url)
-        assert LogMixin.logs
 
 
 @given(st.just(""))  # This is the only case where name is empty
